+++ core/cli/licenses/jopt-simple-5.0.2.jar.sha1
