+++ core/cli/src/main/java/org/elasticsearch/cli/ExitCodes.java
