+++ core/cli/src/main/java/org/elasticsearch/cli/MultiCommand.java
