+++ core/cli/src/main/java/org/elasticsearch/cli/UserException.java
