+++ qa/evil-tests/src/test/java/org/elasticsearch/cli/EvilCommandTests.java
@@ -49,11 +49,11 @@ public class EvilCommandTests extends ESTestCase {
         };
         final MockTerminal terminal = new MockTerminal();
         command.main(new String[0], terminal);
-        assertNotNull(command.shutdownHookThread.get());
+        assertNotNull(command.getShutdownHookThread());
         // successful removal here asserts that the runtime hook was installed in Command#main
-        assertTrue(Runtime.getRuntime().removeShutdownHook(command.shutdownHookThread.get()));
-        command.shutdownHookThread.get().run();
-        command.shutdownHookThread.get().join();
+        assertTrue(Runtime.getRuntime().removeShutdownHook(command.getShutdownHookThread()));
+        command.getShutdownHookThread().run();
+        command.getShutdownHookThread().join();
         assertTrue(closed.get());
         final String output = terminal.getOutput();
         if (shouldThrow) {
